refactor player, inventory types
